requirements-core.txt
Django==5.2.6
psycopg2-binary>=2.9.9,<3.0
django-environ>=0.11.2,<1.0
whitenoise>=6.6.0,<7.0
gunicorn>=21.2.0,<22.0
reportlab>=4.0.4,<5.0
openpyxl>=3.1.2,<4.0

# Machine Learning & Data Science
scikit-learn>=1.3.0,<2.0
pandas>=2.0.0,<3.0
joblib>=1.3.0,<2.0
numpy>=1.24.0,<2.0
matplotlib>=3.7.0,<4.0
seaborn>=0.12.0,<1.0

# Sécurité et Validation
python-magic>=0.4.27,<1.0
django-ratelimit>=4.1.0,<5.0
bleach>=6.1.0,<7.0

# Qualité du code
black>=23.0.0,<24.0
flake8>=6.0.0,<7.0
mypy>=1.5.0,<2.0

# Tests et Couverture
pytest>=7.4.0,<8.0
pytest-django>=4.5.0,<5.0
pytest-cov>=4.1.0,<5.0
coverage>=7.3.0,<8.0
factory-boy>=3.3.0,<4.0
faker>=19.0.0,<20.0
